Adjust parsing tests.
