+++ shardingsphere-test/shardingsphere-parser-test/src/main/java/org/apache/shardingsphere/test/sql/parser/parameterized/asserts/statement/distsql/ral/RALStatementAssert.java
@@ -19,16 +19,16 @@ package org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statemen
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.apache.shardingsphere.distsql.parser.statement.ral.AdvancedDistSQLStatement;
-import org.apache.shardingsphere.distsql.parser.statement.ral.CommonDistSQLStatement;
 import org.apache.shardingsphere.distsql.parser.statement.ral.QueryableRALStatement;
 import org.apache.shardingsphere.distsql.parser.statement.ral.RALStatement;
 import org.apache.shardingsphere.distsql.parser.statement.ral.UpdatableRALStatement;
+import org.apache.shardingsphere.distsql.parser.statement.ral.scaling.QueryableScalingRALStatement;
+import org.apache.shardingsphere.distsql.parser.statement.ral.scaling.UpdatableScalingRALStatement;
 import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.SQLCaseAssertContext;
-import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.AdvancedDistSQLStatementAssert;
-import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.CommonDistSQLStatementAssert;
 import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.QueryableRALStatementAssert;
+import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.scaling.QueryableScalingRALStatementAssert;
 import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.UpdatableRALStatementAssert;
+import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.scaling.UpdatableScalingRALStatementAssert;
 import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.statement.SQLParserTestCase;
 
 /**
@@ -45,14 +45,14 @@ public final class RALStatementAssert {
      * @param expected expected RAL statement test case
      */
     public static void assertIs(final SQLCaseAssertContext assertContext, final RALStatement actual, final SQLParserTestCase expected) {
-        if (actual instanceof QueryableRALStatement) {
+        if (actual instanceof QueryableScalingRALStatement) {
+            QueryableScalingRALStatementAssert.assertIs(assertContext, (QueryableScalingRALStatement) actual, expected);
+        } else if (actual instanceof UpdatableScalingRALStatement) {
+            UpdatableScalingRALStatementAssert.assertIs(assertContext, (UpdatableScalingRALStatement) actual, expected);
+        } else if (actual instanceof QueryableRALStatement) {
             QueryableRALStatementAssert.assertIs(assertContext, (QueryableRALStatement) actual, expected);
         } else if (actual instanceof UpdatableRALStatement) {
             UpdatableRALStatementAssert.assertIs(assertContext, (UpdatableRALStatement) actual, expected);
-        } else if (actual instanceof CommonDistSQLStatement) {
-            CommonDistSQLStatementAssert.assertIs(assertContext, (CommonDistSQLStatement) actual, expected);
-        } else if (actual instanceof AdvancedDistSQLStatement) {
-            AdvancedDistSQLStatementAssert.assertIs(assertContext, (AdvancedDistSQLStatement) actual, expected);
         }
     }
 }
