+++ shardingsphere-test/shardingsphere-parser-test/src/main/java/org/apache/shardingsphere/test/sql/parser/parameterized/asserts/statement/distsql/ral/impl/scaling/query/CheckScalingStatementAssert.java
@@ -15,12 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.update;
+package org.apache.shardingsphere.test.sql.parser.parameterized.asserts.statement.distsql.ral.impl.scaling.query;
 
 import org.apache.commons.collections4.CollectionUtils;
-import org.apache.shardingsphere.scaling.distsql.statement.RestoreScalingSourceWritingStatement;
+import org.apache.shardingsphere.distsql.parser.segment.AlgorithmSegment;
+import org.apache.shardingsphere.scaling.distsql.statement.CheckScalingStatement;
 import org.apache.shardingsphere.test.sql.parser.parameterized.asserts.SQLCaseAssertContext;
-import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.statement.distsql.ral.scaling.RestoreScalingSourceWritingStatementTestCase;
+import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.distsql.ExpectedAlgorithm;
+import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.statement.distsql.ral.scaling.CheckScalingStatementTestCase;
 
 import java.util.List;
 
@@ -30,23 +32,24 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 
 /**
- * Restore scaling source writing statement assert.
+ * Check scaling statement assert.
  */
-public final class RestoreScalingSourceWritingStatementAssert {
+public final class CheckScalingStatementAssert {
     
     /**
-     * Assert restore scaling source writing statement is correct with expected parser result.
+     * Assert check scaling statement is correct with expected parser result.
      *
      * @param assertContext assert context
-     * @param actual actual restore scaling source writing statement
-     * @param expected expected restore scaling source writing statement test case
+     * @param actual actual check scaling statement
+     * @param expected expected check scaling statement test case
      */
-    public static void assertIs(final SQLCaseAssertContext assertContext, final RestoreScalingSourceWritingStatement actual, final RestoreScalingSourceWritingStatementTestCase expected) {
+    public static void assertIs(final SQLCaseAssertContext assertContext, final CheckScalingStatement actual, final CheckScalingStatementTestCase expected) {
         if (null == expected) {
             assertNull(assertContext.getText("Actual statement should not exist."), actual);
         } else {
             assertNotNull(assertContext.getText("Actual statement should exist."), actual);
             assertJobIds(assertContext, actual.getJobId(), expected.getJobIds());
+            assertTypeStrategy(assertContext, actual.getTypeStrategy(), expected.getTableStrategies());
         }
     }
     
@@ -56,6 +59,15 @@ public final class RestoreScalingSourceWritingStatementAssert {
         } else {
             assertNotNull(assertContext.getText("Actual job id should exist."), actual);
             assertThat(assertContext.getText("Job id assertion error"), actual, is(expected.iterator().next()));
+        }
+    }
+    
+    private static void assertTypeStrategy(final SQLCaseAssertContext assertContext, final AlgorithmSegment actual, final List<ExpectedAlgorithm> expected) {
+        if (CollectionUtils.isEmpty(expected)) {
+            assertNull(assertContext.getText("Actual type strategy should not exist."), actual);
+        } else {
+            assertNotNull(assertContext.getText("Actual type strategy should exist."), actual);
+            assertThat(assertContext.getText("Type strategy assertion error"), actual.getName(), is(expected.iterator().next().getName()));
         }
     }
 }
